python/promots_examples/requirements.txt
python-gflags
openai
py-dotenv
